requirements.txt
numpy 
pandas
streamlit
nltk
SpeechRecognition
pydub
gTTs
tensorflow
Pillow
ipython
protobuf==3.20.1
scikit-learn